Fix data extraction

+++ src/Modules/CrestApps.OrchardCore.AI/Indexes/AIChatSessionIndexProvider.cs
@@ -27,7 +27,7 @@ internal sealed class AIChatSessionIndexProvider : IndexProvider<AIChatSession>
                     ClientId = session.ClientId,
                     CreatedUtc = session.CreatedUtc,
                     Title = Str.Truncate(session.Title, 255),
-                    Status = (int)session.Status,
+                    Status = session.Status,
                     LastActivityUtc = session.LastActivityUtc,
                 };
             });
